migrate __getitem__ from Variable to NamedArray
